+++ saleor/checkout/checkout_cleaner.py
@@ -1,12 +1,13 @@
-from typing import Iterable, Type, Union
+from typing import Iterable, Optional, Type, Union
 
 from django.core.exceptions import ValidationError
 
-from ...checkout.error_codes import CheckoutErrorCode
-from ...checkout.models import Checkout, CheckoutLine
-from ...checkout.utils import is_fully_paid, is_valid_shipping_method
-from ...discount import DiscountInfo
-from ...payment.error_codes import PaymentErrorCode
+from ..discount import DiscountInfo
+from ..payment import gateway, models as payment_models
+from ..payment.error_codes import PaymentErrorCode
+from .error_codes import CheckoutErrorCode
+from .models import Checkout, CheckoutLine
+from .utils import is_fully_paid, is_valid_shipping_method
 
 
 def clean_checkout_shipping(
@@ -65,9 +66,11 @@ def clean_checkout_payment(
     lines: Iterable[CheckoutLine],
     discounts: Iterable[DiscountInfo],
     error_code: Type[CheckoutErrorCode],
+    last_payment: Optional[payment_models.Payment],
 ):
     clean_billing_address(checkout, error_code)
     if not is_fully_paid(checkout, lines, discounts):
+        gateway.payment_refund_or_void(last_payment)
         raise ValidationError(
             "Provided payment methods can not cover the checkout's total amount",
             code=error_code.CHECKOUT_NOT_FULLY_PAID.value,
